add dummy middleware for authentication
